pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sprouts-coupons"
version = "0.0.1"
description = "clip sprouts coupons"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Brendan King", email = "kingbrendan94@gmail.com" },
]
keywords = ["sprouts, claude"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "dacite~=1.8.1",
    "playwright>=1.40.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "dictdiffer~=0.9.0",
    "pytest>=7.0.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
sprouts-coupons = "sprouts_coupons.main:main"

[project.urls]
Homepage = "https://github.com/kingb12/sprouts-coupons"

[tool.hatch.build.targets.wheel]
packages = ["src/sprouts_coupons"]

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit_build: marks tests as unit build tests (run with: pytest -m unit_build)",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow-running",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports = "normal"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["sprouts_coupons"]